controller method returning list of mcgill buildings (the rest of the buildings will be added later)

+++ mcgill-augmented-reality-backend/src/main/java/mcgillar/backend/services/location/LocationService.java
@@ -12,7 +12,7 @@ public class LocationService {
     
     private LocationRepository locationRepository;
     
-    public Location createLocation(String postalCode, String address, Long latitude, Long longitude, Long altitude) {
+    public Location createLocation(String postalCode, String address, Double latitude, Double longitude, Double altitude) {
         if (latitude == null || longitude == null) return null;
         Location location = new Location(postalCode, address, latitude, longitude, altitude);
         return locationRepository.save(location);
